+++ src/render/providers/skqw-utils.ts
@@ -1,0 +1,8 @@
+
+function foo() {
+    console.log('bar!');
+}
+
+export const skqwUtils = {
+    foo
+};